hoist stencil pattern selection in TBB impl
had to make tbb_partitioner global variable because i have not yet
figured out how to declare a proper C++11 function pointer equivalent
and auto can't deduce this for a template function.  given that
tbb_partitioner is supposed to persist anyways, global scope is not as
evil as it is in other cases.  well, it was already global, so yeah,
nothing worse here...