detect ancient libstdc++ to improve user experience
Intel C++ claims it doesn't support C++11 when it most certainly does if
the underlying GCC/libstdc++ are ancient (e.g. GCC 4.4).

Before we tell the user that C++11 is not supported, verify that the
standard library supports the C++11 ABI, which we hope is consistent
with said library supporting C++11.