Cilk is deprecated.  Remove C++ implementation.
Cilk is deprecated as of the Intel 18 compiler release [1].
This removes the C++/Cilk implementations of Stencil and Transpose.
This also switches "pragma simd" to "pragma vector".

Intel recommends "pragma omp simd" as the replacement for "pragma simd",
which might be better, but that requires more preprocessor logic and
will be postponed to a future commit.

We will leave the C/Cilk implementation for now, since it may be useful
in the context of CPLEX (ISO C parallelism, not the IBM thing) activities.

Note also that C++/Cilk with the Intel compiler has had major
performance issues in Stencil, which appear to be related to the C++ optimizer
giving up.  This issue does not appear in the C/Cilk implementation, so
even if Cilk was supported, implementation quality favors deleting the
C++/Cilk code.

[1] https://software.intel.com/en-us/articles/migration-from-intelr-cilktm-plus-to-openmp-or-intelr-threading-building-blocks-intelr-tbb