apply scalar optimizations to p2p
Intel Fortran turns array references into scalars.  Intel C/C++ will do
it for 2D arrays but that's tricky here without VLAs.  The optimization
can be performed by hand, which is what this commit is.

Thanks to Martyn Corden of Intel for analyzing this issue and providing
the improved implementation.