switch to LLVM PSTL for Travis and other cleanup
Intel PSTL was not working with LLVM 9 so we switch to the LLVM PSTL
repo for Travis testing.

We shifted to raw duplication for STL and PSTL to eliminate preprocessor
workarounds.  This should be undone later by refactoring the for_each
and such into the PRK namespace where the optional execution policy can
be hidden.

Signed-off-by: Jeff Hammond <jeff.r.hammond@intel.com>