remove clunky device detection code
previously, we tried to detect device presence with preprocessor macros
and querying for the SPIR extension (because that worked for
ComputeCpp).  because Intel Level Zero does not support that method, we
cannot use it in DPC++ without adding even more clunkiness.

this removes SPIR/PTX queries and uses try-catch for every device in all
SYCL implementations.  if we find implementations that do not throw
properly when devices do not work, then that is on them (we will file
bug reports) but we can work around that other ways that are yet to be
determined.

Signed-off-by: Jeff Hammond <jeff.r.hammond@intel.com>