add block_size runtime selection for nstream
DPC++ block size default for CUDA was 32, which was slow
(~650 GB/s vs ~840 GB/s)

This change uses nd_range in the DPC++ version (it should also be applied
to the SYCL code), sets the default to 256 like the CUDA version, makes
the block_sizes in both DPC++ and CUDA runtime options, and sets the
default for CPU DPC++ to a big number because that is good.

Thanks to James Brodman for helping me figure this out.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>