perf test measurement
we do this
B+=transpose(A)
A+=1
which does read(A), read(B), store(B), store(A), and possibly a second read(A) if the loops are split.
we measure bandwidth with 2*bytes, but this doesn't make sense to me.  shouldn't it be 4x?
if we remove A+=1, we have read(A), read(B), write(B), which is 3x, or 1.5x if we use the prior convention.